logs.py: don't print (unchanged) lines with --no-log unless DEBUG.
This accidentally made output look different with --no-log vs with
normal redo-log output, because redo-log has a -u option while plain
redo does not.

(This is on purpose.  When running redo, you only want to see the things
that actually happened, so it never passes -u to the auto-launched
redo-log instance.  But when reviewing logs later, you might want to
look at the past logs from building a component that was unchanged in
the most recent run.)