shelltest.od: warning 84 (W84) triggered on *all* shells, not just posh.
I must have changed this at the last minute when adding it, but I don't
know why.  We were redefining f4() inside a subshell, so it never
applied to the parent shell, so it was always considered a failure.
But that's not what we were supposed to be testing.

This is just supposed to be a test of the really rare syntax of
defining a function without enclosing braces, which only fails on posh,
as far as I've seen.