Prevent "Exception ... ignored" in `redo-log ... | head`
When STDOUT is piped to another program and that program closes the pipe,
this program is supposed to terminate. However, Python 3 prints an error
message because prior to exiting, STDOUT is automatically flushed which
does not work due to it being closed by the other side.

This commit includes code taken from the Python documentation for this case.
The output stream is redirected to /dev/null