minimal/do: allow for relative shebangs
The current version of do expects #! to be followed by a /. A variety of
tested systems treat the first word which follows #! as a path (which
could be absolute or relative). Additionally, #! can be followed by
spaces before the first path character. This patch addresses this
potential incompatibility as a side effect.

An example usecase of relative paths in shebang in a redo based build
system can be seen here:

https://git.sr.ht/~tk/pack/commit/01438f00d44adf33f59da729b347eca60cb461a8

This patch was tested to work in a variety of situations and systems.